// src/models/document.model.ts
import mongoose, { Schema, model, Document } from 'mongoose';

export interface IDocument extends Document {
    userId: mongoose.Types.ObjectId,
    fileName: string;
    textChunk: string;         // The actual readable paragraph
    embedding: number[];       // The mathematical vector array generated by Gemini
    createdAt: Date;
}

const documentSchema = new Schema<IDocument>(
  {
    userId: {
        type: Schema.Types.ObjectId,
        ref: 'User',
        required: [true, 'A document chunk must be bound to a valid user account.']
    },
    fileName: { 
      type: String, 
      required: [true, 'A document must track its source file name'],
      trim: true 
    },
    textChunk: { 
      type: String, 
      required: [true, 'Content chunk cannot be empty'] 
    },
    embedding: { 
      type: [Number], 
      required: [true, 'Vector embedding is required for semantic search'] 
    }
  },
  { timestamps: { createdAt: true, updatedAt: false } } // We don't need updatedAt for static knowledge chunks
);

export const KnowledgeDocument = model<IDocument>('KnowledgeDocument', documentSchema);